+++ NEWS.md
@@ -1,6 +1,8 @@
 # mulgar 1.0.1.9000
 
 * Fix with is_tibble
+* Added function to make a slice for a single projection
+* Added choice of number of PCs in ggscree
 
 # mulgar 1.0.1